Refactoring components utils
