+++ src/main/java/org/istic/synthlab/components/eg/Eg.java
@@ -1,6 +1,6 @@
 package org.istic.synthlab.components.eg;
 
-import org.istic.synthlab.core.AbstractComponent;
+import org.istic.synthlab.components.AbstractComponent;
 import org.istic.synthlab.core.modules.envelope.IEnvelopeDAHDSR;
 import org.istic.synthlab.core.modules.io.IInput;
 import org.istic.synthlab.core.modules.io.IOutput;
