+++ src/main/java/org/istic/synthlab/components/generic/generic.java
@@ -1,8 +1,6 @@
 package org.istic.synthlab.components.generic;
 
-import org.istic.synthlab.core.AbstractComponent;
-
-import javax.swing.*;
+import org.istic.synthlab.components.AbstractComponent;
 
 /**
  * Created by seb on 04/02/16.
