+++ src/main/java/org/istic/synthlab/components/vcoa/Vcoa.java
@@ -1,6 +1,6 @@
 package org.istic.synthlab.components.vcoa;
 
-import org.istic.synthlab.core.AbstractComponent;
+import org.istic.synthlab.components.AbstractComponent;
 import org.istic.synthlab.core.modules.io.IOutput;
 import org.istic.synthlab.core.modules.modulators.IModulator;
 import org.istic.synthlab.core.modules.modulators.ModulatorType;
