+++ src/main/java/org/istic/synthlab/core/modules/io/Input.java
@@ -1,7 +1,7 @@
 package org.istic.synthlab.core.modules.io;
 
 import com.jsyn.ports.UnitInputPort;
-import org.istic.synthlab.core.IComponent;
+import org.istic.synthlab.components.IComponent;
 import org.istic.synthlab.core.services.Register;
 
 /**
