+++ src/main/java/org/istic/synthlab/core/modules/modulators/AbstractModulator.java
@@ -1,6 +1,6 @@
 package org.istic.synthlab.core.modules.modulators;
 
-import org.istic.synthlab.core.IComponent;
+import org.istic.synthlab.components.IComponent;
 import org.istic.synthlab.core.modules.io.IInput;
 import org.istic.synthlab.core.modules.io.IOutput;
 import org.istic.synthlab.core.utils.parametrization.Potentiometer;
