+++ src/main/java/org/istic/synthlab/core/modules/modulators/VcaAmplitudeModulator.java
@@ -1,6 +1,6 @@
 package org.istic.synthlab.core.modules.modulators;
 
-import org.istic.synthlab.core.IComponent;
+import org.istic.synthlab.components.IComponent;
 import org.istic.synthlab.core.services.Factory;
 import org.istic.synthlab.core.services.Register;
 import org.istic.synthlab.core.utils.jsyn.VcaFunction;
