+++ src/main/java/org/istic/synthlab/core/modules/oscillators/RedNoiseOscillator.java
@@ -1,7 +1,7 @@
 package org.istic.synthlab.core.modules.oscillators;
 
 import com.jsyn.unitgen.RedNoise;
-import org.istic.synthlab.core.IComponent;
+import org.istic.synthlab.components.IComponent;
 
 /**
  * A triangle shape generator
