+++ src/main/java/org/istic/synthlab/core/modules/passThrough/PassThrough.java
@@ -1,6 +1,6 @@
 package org.istic.synthlab.core.modules.passThrough;
 
-import org.istic.synthlab.core.IComponent;
+import org.istic.synthlab.components.IComponent;
 import org.istic.synthlab.core.modules.io.IInput;
 import org.istic.synthlab.core.modules.io.IOutput;
 import org.istic.synthlab.core.services.Factory;
