+++ src/main/java/org/istic/synthlab/core/modules/visualizer/Visualizer.java
@@ -1,7 +1,7 @@
 package org.istic.synthlab.core.modules.visualizer;
 
 import com.jsyn.scope.AudioScope;
-import org.istic.synthlab.core.IComponent;
+import org.istic.synthlab.components.IComponent;
 import org.istic.synthlab.core.modules.io.IOutput;
 import org.istic.synthlab.core.services.Factory;
 import org.istic.synthlab.core.services.Register;
