+++ src/test/java/org/istic/synthlab/components/oscillators/SineOscillatorTest.java
@@ -3,7 +3,7 @@ package org.istic.synthlab.components.oscillators;
 import com.jsyn.unitgen.UnitOscillator;
 import junit.framework.TestCase;
 import org.istic.synthlab.components.vcoa.Vcoa;
-import org.istic.synthlab.core.IComponent;
+import org.istic.synthlab.components.IComponent;
 import org.istic.synthlab.core.modules.io.IInput;
 import org.istic.synthlab.core.modules.io.IOutput;
 import org.istic.synthlab.core.modules.oscillators.SineOscillator;
