+++ src/test/java/org/istic/synthlab/core/ChannelTest.java
@@ -5,6 +5,7 @@ import com.jsyn.ports.UnitInputPort;
 import com.jsyn.ports.UnitOutputPort;
 import javafx.util.Pair;
 import junit.framework.TestCase;
+import org.istic.synthlab.components.IComponent;
 import org.istic.synthlab.components.vcoa.Vcoa;
 import org.istic.synthlab.core.modules.io.IInput;
 import org.istic.synthlab.core.modules.io.IOutput;
