+++ src/test/java/org/istic/synthlab/core/io/InputTest.java
@@ -7,7 +7,7 @@ import junit.framework.TestCase;
 import org.istic.synthlab.components.out.Out;
 import org.istic.synthlab.components.vcoa.Vcoa;
 import org.istic.synthlab.core.Channel;
-import org.istic.synthlab.core.IComponent;
+import org.istic.synthlab.components.IComponent;
 import org.istic.synthlab.core.modules.io.IInput;
 import org.istic.synthlab.core.modules.io.IOutput;
 import org.istic.synthlab.core.modules.io.Input;
