+++ src/test/java/org/istic/synthlab/core/modules/envelope/EnvelopeDAHDSRTest.java
@@ -5,7 +5,7 @@ import junit.framework.TestCase;
 import org.istic.synthlab.components.eg.Eg;
 import org.istic.synthlab.components.out.Out;
 import org.istic.synthlab.components.vcoa.Vcoa;
-import org.istic.synthlab.core.IComponent;
+import org.istic.synthlab.components.IComponent;
 import org.istic.synthlab.core.modules.lineOuts.ILineOut;
 import org.istic.synthlab.core.modules.lineOuts.LineType;
 import org.istic.synthlab.core.modules.oscillators.IOscillator;
