+++ src/test/java/org/istic/synthlab/core/modules/oscillators/SineOscillatorTest.java
@@ -3,17 +3,15 @@ package org.istic.synthlab.core.modules.oscillators;
 import com.jsyn.Synthesizer;
 import org.istic.synthlab.components.out.Out;
 import org.istic.synthlab.components.replicator.Replicator;
-import org.istic.synthlab.core.IComponent;
+import org.istic.synthlab.components.IComponent;
 import org.istic.synthlab.core.modules.lineOuts.ILineOut;
 import org.istic.synthlab.core.modules.lineOuts.LineType;
-import org.istic.synthlab.core.modules.passThrough.PassThrough;
 import org.istic.synthlab.core.services.Factory;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mockito;
-import org.mockito.internal.exceptions.ExceptionIncludingMockitoWarnings;
 import org.mockito.runners.MockitoJUnitRunner;
 
 import static org.junit.Assert.assertNotSame;
